Setup abbreviation behavior to be like an isk-less Vim.
For now, the internal equivalent of "iskeyword" is [^\s'"].

In the future, the internal "iskeyword" equivalent can be expanded. It may
be overboard to add an "iskeyword" option, as it applies to lots more
things than :abbr. Additionally, it's not clear why it's bad to call
keyword characters everything except whitespace and quotes.

TODO: Should abbreviations be triggered by any non-keyword character?

TODO: Should abbreviations be triggered by <CR>?

TODO: Should abbreviations be able to include <Left>, etc.?