Added some instructions to vimperator.vim.
Files like vimperator.vim, javascript.vim, and vimperatorrc.example
might fit better in a "contrib" directory. It's not clear whether there
should be vimperator/contrib and muttator/contrib directories or a
single liberator/contrib directory.

The Makefile.common doesn't include these files in the bundle, and so it
doesn't seem like a bad idea to package them in a single
liberator/contrib directory with a CONTENTS or README file giving a ToC
for the directory.